Add .DS_Store to .gitignore files in multiple directories
This commit adds ".DS_Store" to the .gitignore files across several directories. The ".DS_Store" file is a system file that is automatically generated by macOS and should not be tracked in git. Accordingly, by including it in the .gitignore file, unwanted changes and clutter will be eliminated from future commits.